Moved out the Use Case Request and Response objects into the Boundaries folder. Adjusted the namespaces.
