+++ property-api/V1/UseCase/GetMultipleProperties/IGetMultiplePropertiesUseCase.cs
@@ -1,4 +1,5 @@
 using property_api.V1.Domain;
+using property_api.V1.UseCase.GetMultipleProperties.Boundaries;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,15 +10,5 @@ namespace property_api.V1.UseCase.GetMultipleProperties
     public interface IGetMultiplePropertiesUseCase 
     {
         GetMultiplePropertiesUseCaseResponse Execute(GetMultiplePropertiesUseCaseRequest request);
-    }
-
-    public class GetMultiplePropertiesUseCaseRequest
-    {
-        public IList<string> PropertyRefs { get;set;}
-    }
-
-    public class GetMultiplePropertiesUseCaseResponse
-    {
-        public IList<Property> Properties { get;set;}
     }
 }
