+++ property-api/V1/UseCase/GetMultipleProperties/Impl/GetMultiplePropertiesUseCase.cs
@@ -1,11 +1,12 @@
 using property_api.V1.Domain;
 using property_api.V1.Gateways.GetMultipleProperties;
+using property_api.V1.UseCase.GetMultipleProperties.Boundaries;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace property_api.V1.UseCase.GetMultipleProperties
+namespace property_api.V1.UseCase.GetMultipleProperties.Impl
 {
     public class GetMultiplePropertiesUseCase : IGetMultiplePropertiesUseCase
     {
